+++ src/lib/mymathlib.com/singular_value_decomposition.h
@@ -6,15 +6,17 @@
   extern "C" {
 #endif
 
-int Singular_Value_Decomposition(float* A, int nrows, int ncols, float* U,
+#include <stdint.h>
+
+int8_t Singular_Value_Decomposition(float* A, int8_t nrows, int8_t ncols, float* U,
                       float* singular_values, float* V, float* dummy_array);
 
 
 void Singular_Value_Decomposition_Solve(float* U, float* D, float* V,
-                float tolerance, int nrows, int ncols, float *B, float* x);
+                float tolerance, int8_t nrows, int8_t ncols, float *B, float* x);
 
 
 void Singular_Value_Decomposition_Inverse(float* U, float* D, float* V,
-                        float tolerance, int nrows, int ncols, float *Astar);
+                        float tolerance, int8_t nrows, int8_t ncols, float *Astar);
 
 #endif
